#4 Added ability to run docker-compose builds on digitalocean
