+++ compose-dev.yml
@@ -1,0 +1,6 @@
+version: '3'
+services:
+
+  backend-nodejs:
+    volumes:
+     - ./servercode/src:/src
