+++ servercode/package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "prestart": "scripts/docker-start.sh",
     "test": "mocha src/**/**.spec.js",
-    "setup": "scripts/docker-run.sh"
+    "setup": "scripts/docker-run.sh",
+    "start": "nodemon src/index.js"
   },
   "author": "Artur Dzidzoiev",
   "license": "ISC",
@@ -13,6 +14,7 @@
   "dependencies": {
     "app-module-path": "^2.2.0",
     "body-parser": "^1.17.1",
+    "consul": "^0.29.0",
     "express": "^4.15.2",
     "mongodb": "^2.2.24",
     "mongoose": "^4.9.1",
@@ -23,7 +25,6 @@
     "assert": "^1.4.1",
     "chai": "^3.5.0",
     "mocha": "^3.2.0",
-    "nodemon": "^1.11.0",
     "sinon": "^2.1.0"
   }
 }
