--- servercode/server.js
@@ -1,5 +1,0 @@
-require('app-module-path').addPath(__dirname + '/src');
-
-const index = require('./src/index');
-
-index.run();
