+++ servercode/src/index.js
@@ -1,20 +1,5 @@
-const express = require("express");
-const config = require("config/config");
-const morgan = require("morgan");
-const mongo = require('util/mongo');
-const bodyParser = require('body-parser');
+require('app-module-path').addPath(__dirname);
 
-const app = express();
+const server = require('./server');
 
-app.use(require('./cors'));
-app.use(morgan('tiny'));
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
-app.use('/api', require('router'));
-
-exports.run = () => {
-    mongo.connect()
-        .then(() => app.listen(config.express.port))
-        .then(() => console.warn("Listening to http://localhost:" + config.express.port));
-};
-
+server.run();
